86czvr7pd: Refactor add task modal into a separate file
